+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Playlists can now be shuffled as they are added to the queue, using the `shuffle` option to `play`.
+
 ### Fixed
 - The SQLite database path is now correctly generated on Windows